Refactor: 객실 리스트 props 전달 방식 수정
