+++ src/components/Accomodation/AccomodationDetail/AccomodationDetail.tsx
@@ -1,5 +1,6 @@
+import * as React from "react";
 import AccomodationInfo from "../AccomodationInfo";
-import AccomodationRoomList from "../AccomodationRoomList";
+import AccomodationRoomList, { RoomProps } from "../AccomodationRoomList";
 import AccomodationSelect from "../AccomodationSelect";
 import { ImageStyled, ImageBox } from "./AccomodationDetail.styles";
 import thumb from "/public/images/thumb.jpg";
@@ -17,7 +18,46 @@ const AccomodationInfoData = {
   cookingInfo: "객실 내 취사 가능",
 };
 
-const AccomodationDetail = () => {
+const rooms: RoomProps[] = [
+  {
+    id: 1,
+    name: "슈페리어 트윈 1",
+    maxPerson: 4,
+    minPerson: 2,
+    extraPrice: 20000,
+    totalPrice: 142000,
+    image: thumb,
+  },
+  {
+    id: 2,
+    name: "슈페리어 트윈 2",
+    maxPerson: 4,
+    minPerson: 2,
+    extraPrice: 20000,
+    totalPrice: 142000,
+    image: thumb,
+  },
+  {
+    id: 3,
+    name: "슈페리어 트윈 3",
+    maxPerson: 4,
+    minPerson: 2,
+    extraPrice: 20000,
+    totalPrice: 142000,
+    image: thumb,
+  },
+  {
+    id: 4,
+    name: "슈페리어 트윈 4",
+    maxPerson: 4,
+    minPerson: 2,
+    extraPrice: 20000,
+    totalPrice: 142000,
+    image: thumb,
+  },
+];
+
+const AccomodationDetail: React.FC = () => {
   return (
     <Container>
       <ImageBox>
@@ -25,7 +65,7 @@ const AccomodationDetail = () => {
       </ImageBox>
       <AccomodationInfo {...AccomodationInfoData} />
       <AccomodationSelect />
-      <AccomodationRoomList />
+      <AccomodationRoomList rooms={rooms} />
     </Container>
   );
 };
